jitter: expose pink-noise synthesis as public sims module
- Library consumers (e.g. tracking-test-bench scene-camera) can now build
  jitter trajectories without copy-pasting from the CLI binary.
- Adds spectral-slope unit test confirming Kellett's 7-pole filter
  approximates -10 dB/decade over its validated band.

Closes #93
